+++ gitlab/__init__.py
@@ -118,7 +118,6 @@ class Gitlab(object):
         self.users = objects.UserManager(self)
         self.todos = objects.TodoManager(self)
         if self._api_version == '3':
-            self.keys = objects.KeyManager(self)
             self.teams = objects.TeamManager(self)
         else:
             self.dockerfiles = objects.DockerfileManager(self)
@@ -198,12 +197,6 @@ class Gitlab(object):
         else:
             self._credentials_auth()
 
-    def credentials_auth(self):
-        """Performs an authentication using email/password."""
-        warnings.warn('credentials_auth() is deprecated and will be removed.',
-                      DeprecationWarning)
-        self._credentials_auth()
-
     def _credentials_auth(self):
         if not self.email or not self.password:
             raise GitlabAuthenticationError("Missing email/password")
@@ -220,12 +213,6 @@ class Gitlab(object):
             self.user = self._objects.CurrentUser(manager, r)
 
         self._set_token(self.user.private_token)
-
-    def token_auth(self):
-        """Performs an authentication using the private token."""
-        warnings.warn('token_auth() is deprecated and will be removed.',
-                      DeprecationWarning)
-        self._token_auth()
 
     def _token_auth(self):
         if self.api_version == '3':
@@ -256,17 +243,6 @@ class Gitlab(object):
 
         return self._server_version, self._server_revision
 
-    def set_url(self, url):
-        """Updates the GitLab URL.
-
-        Args:
-            url (str): Base URL of the GitLab server.
-        """
-        warnings.warn('set_url() is deprecated, create a new Gitlab instance '
-                      'if you need an updated URL.',
-                      DeprecationWarning)
-        self._url = '%s/api/v%s' % (url, self._api_version)
-
     def _construct_url(self, id_, obj, parameters, action=None):
         if 'next_url' in parameters:
             return parameters['next_url']
@@ -291,17 +267,6 @@ class Gitlab(object):
         else:
             return url
 
-    def set_token(self, token):
-        """Sets the private token for authentication.
-
-        Args:
-            token (str): The private token.
-        """
-        warnings.warn('set_token() is deprecated, use the private_token '
-                      'argument of the Gitlab constructor.',
-                      DeprecationWarning)
-        self._set_token(token)
-
     def _set_token(self, private_token, oauth_token=None):
         self.private_token = private_token if private_token else None
         self.oauth_token = oauth_token if oauth_token else None
@@ -314,19 +279,6 @@ class Gitlab(object):
             self.headers['Authorization'] = "Bearer %s" % oauth_token
             if "PRIVATE-TOKEN" in self.headers:
                 del self.headers["PRIVATE-TOKEN"]
-
-    def set_credentials(self, email, password):
-        """Sets the email/login and password for authentication.
-
-        Args:
-            email (str): The user email or login.
-            password (str): The user password.
-        """
-        warnings.warn('set_credentials() is deprecated, use the email and '
-                      'password arguments of the Gitlab constructor.',
-                      DeprecationWarning)
-        self.email = email
-        self.password = password
 
     def enable_debug(self):
         import logging