+++ gitlab/v3/objects.py
@@ -19,7 +19,6 @@ from __future__ import print_function
 from __future__ import absolute_import
 import base64
 import json
-import warnings
 
 import six
 from six.moves import urllib
@@ -293,23 +292,6 @@ class BroadcastMessage(GitlabObject):
 
 class BroadcastMessageManager(BaseManager):
     obj_cls = BroadcastMessage
-
-
-class Key(GitlabObject):
-    _url = '/deploy_keys'
-    canGet = 'from_list'
-    canCreate = False
-    canUpdate = False
-    canDelete = False
-
-    def __init__(self, *args, **kwargs):
-        warnings.warn("`Key` is deprecated, use `DeployKey` instead",
-                      DeprecationWarning)
-        super(Key, self).__init__(*args, **kwargs)
-
-
-class KeyManager(BaseManager):
-    obj_cls = Key
 
 
 class DeployKey(GitlabObject):
@@ -2043,11 +2025,6 @@ class Project(GitlabObject):
         raise_error_from_response(r, GitlabCreateError, 201)
         return Project(self.gitlab, r.json()) if r.status_code == 201 else self
 
-    def archive_(self, **kwargs):
-        warnings.warn("`archive_()` is deprecated, use `archive()` instead",
-                      DeprecationWarning)
-        return self.archive(**kwargs)
-
     def unarchive(self, **kwargs):
         """Unarchive a project.
 
@@ -2062,12 +2039,6 @@ class Project(GitlabObject):
         r = self.gitlab._raw_delete(url, **kwargs)
         raise_error_from_response(r, GitlabCreateError, 201)
         return Project(self.gitlab, r.json()) if r.status_code == 201 else self
-
-    def unarchive_(self, **kwargs):
-        warnings.warn("`unarchive_()` is deprecated, "
-                      "use `unarchive()` instead",
-                      DeprecationWarning)
-        return self.unarchive(**kwargs)
 
     def share(self, group_id, group_access, **kwargs):
         """Share the project with a group.